test ContainerComposite

+++ src/Util/Lists/ObjectPriorityList.php
@@ -17,7 +17,9 @@ class ObjectPriorityList implements IteratorAggregate, Countable
 
     public function add($object, int $priority): void
     {
-        $this->delete($object);
+        if ($this->has($object)) {
+            $this->delete($object);
+        }
         $this->_set($this->objectID($object), $object, $priority);
     }
